refactor(android): remove orphaned legacy screens
