+++ apps/android/style.md
@@ -104,10 +104,10 @@ Hard rule: avoid ultra-thin weights on light backgrounds.
 
 ## 11. Source Of Truth
 
-- `app/src/main/java/ai/openclaw/android/ui/OpenClawTheme.kt`
-- `app/src/main/java/ai/openclaw/android/ui/OnboardingFlow.kt`
-- `app/src/main/java/ai/openclaw/android/ui/RootScreen.kt`
-- `app/src/main/java/ai/openclaw/android/ui/SettingsSheet.kt`
-- `app/src/main/java/ai/openclaw/android/MainViewModel.kt`
+- `app/src/main/java/ai/openclaw/app/ui/OpenClawTheme.kt`
+- `app/src/main/java/ai/openclaw/app/ui/OnboardingFlow.kt`
+- `app/src/main/java/ai/openclaw/app/ui/RootScreen.kt`
+- `app/src/main/java/ai/openclaw/app/ui/SettingsScreens.kt`
+- `app/src/main/java/ai/openclaw/app/MainViewModel.kt`
 
 If style and implementation diverge, update both in the same change.
